requirements.txt
flask
flask-cors
python-dotenv
requests
pytube>=15.0.0
PyPDF2>=2.0.0
python-magic
openai
pydantic
pymongo==4.6.1
motor==3.3.2
sqlalchemy
pdfminer.six